restructure classes to use inheritance
